reduce file size for e2e (#1723)
* reduce file size for e2e
