+++ libs/e2e/src/index.ts
@@ -9,5 +9,5 @@ export * from "./lib/describer/modelAssessment/errorAnalysis/describeErrorAnalys
 export * from "./lib/describer/modelAssessment/featureImportances/individualFeatureImportance/describeIndividualFeatureImportance";
 export * from "./lib/describer/modelAssessment/modelOverview/describeModelOverview";
 export * from "./lib/describer/modelAssessment/whatIfCounterfactuals/describeWhatIf";
-export * from "./lib/describer/modelAssessment/modelAssessmentDatasets";
+export * from "./lib/describer/modelAssessment/datasets/modelAssessmentDatasets";
 export * from "./lib/describer/modelAssessment/IModelAssessmentData";
