+++ libs/e2e/src/lib/describer/modelAssessment/dataExplorer/describeDatasetExplorer.ts
@@ -2,11 +2,11 @@
 // Licensed under the MIT License.
 
 import { Locators } from "../Constants";
+import { modelAssessmentDatasets } from "../datasets/modelAssessmentDatasets";
 import {
   IModelAssessmentData,
   RAINotebookNames
 } from "../IModelAssessmentData";
-import { modelAssessmentDatasets } from "../modelAssessmentDatasets";
 
 import { describeAggregatePlot } from "./describeAggregatePlot";
 import { describeCohortFunctionality } from "./describeCohortFunctionality";
